requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
pytesseract==0.3.10
spacy==3.7.2
transformers==4.34.0
torch==2.1.0
redis==5.0.1
plotly==5.17.0
openai==1.1.1
pandas==2.1.1
numpy==1.26.0
pillow==10.0.1
matplotlib==3.8.0
wordcloud==1.9.2
aiofiles==23.2.1
tenacity==8.2.3